remap.h: changed type of dims to size_t.

+++ src/remap.h
@@ -87,7 +87,7 @@ typedef struct {
   bool     non_global;
   bool     is_cyclic;
 
-  int      dims[2];               /* size of grid dimension */
+  size_t   dims[2];               /* size of grid dimension */
 
   int      nvgp;                  /* size of vgpm           */
   int*     vgpm;                  /* flag which cells are valid   */
@@ -223,11 +223,11 @@ int grid_search_reg2d_nn(size_t nx, size_t ny, size_t *restrict nbr_add, double 
                          const double *restrict src_center_lat, const double *restrict src_center_lon);
 
 int grid_search_reg2d(remapgrid_t *src_grid, size_t *restrict src_add, double *restrict src_lats, 
-                      double *restrict src_lons,  double plat, double plon, const int *restrict src_grid_dims,
+                      double *restrict src_lons,  double plat, double plon, const size_t *restrict src_grid_dims,
                       const double *restrict src_center_lat, const double *restrict src_center_lon);
 
 int grid_search(remapgrid_t *src_grid, size_t *restrict src_add, double *restrict src_lats, 
-		double *restrict src_lons,  double plat, double plon, const int *restrict src_grid_dims,
+		double *restrict src_lons,  double plat, double plon, const size_t *restrict src_grid_dims,
 		const double *restrict src_center_lat, const double *restrict src_center_lon,
 		const restr_t *restrict src_grid_bound_box, const size_t *restrict src_bin_add);
